% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_basic_list.R
\name{make_basic_list}
\alias{make_basic_list}
\title{Create a list of Graph objects}
\usage{
make_basic_list(object, op, filter, n, ...)
}
\arguments{
\item{object}{An R6 object inheriting from \code{AzureGraph::ms_object}.}

\item{op}{A string, the REST operation.}

\item{filter, n}{Filtering arguments for the list.}

\item{...}{Arguments passed to lower-level functions, and ultimately to \code{AzureGraph::call_graph_endpoint}.}
}
\value{
If \code{n} is NULL, an iterator object, of class \code{AzureGraph::ms_graph_pager}. Otherwise, a list of individual Graph objects.
}
\description{
Create a list of Graph objects
}
\details{
This function is a basic utility called by various Microsoft365R class methods. It is exported to work around issues in how R6 handles classes that extend across package boundaries. It should not be called by the user.
}
\seealso{
\code{\link[AzureGraph:call_graph]{AzureGraph::call_graph_endpoint}}, \code{\link[AzureGraph:ms_graph_pager]{AzureGraph::ms_graph_pager}}
}
