% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_channel.R
\docType{class}
\name{ms_channel}
\alias{ms_channel}
\title{Teams channel}
\format{
An R6 object of class \code{ms_channel}, inheriting from \code{ms_object}.
}
\description{
Class representing a Microsoft Teams channel.
}
\section{Fields}{

\itemize{
\item \code{token}: The token used to authenticate with the Graph host.
\item \code{tenant}: The Azure Active Directory tenant for the parent drive.
\item \code{type}: Always "channel" for a channel object
\item \code{team_id}: The ID of the parent team.
\item \code{properties}: The item properties (metadata).
}
}

\section{Methods}{

\itemize{
\item \code{new(...)}: Initialize a new object. Do not call this directly; see 'Initialization' below.
\item \code{delete(confirm=TRUE)}: Delete this channel. By default, ask for confirmation first.
\item \code{update(...)}: Update the channel's properties (metadata) in Microsoft Graph.
\item \code{do_operation(...)}: Carry out an arbitrary operation on the channel.
\item \code{sync_fields()}: Synchronise the R object with the channel metadata in Microsoft Graph.
\item \code{send_message(body, content_type, attachments)}: Sends a new message to the channel.
\item \code{list_messages(n=50)}: Retrieves the messages in the channel. By default, this is limited to the 50 most recent messages; set the \code{n} argument to change this.
\item \code{get_message(message_id)}: Retrieves a specific message in the channel.
\item \code{delete_message(message_id, confirm=TRUE)}: Deletes a message. By default, ask for confirmation first.
\item \code{list_files()}: List the files for the channel.
\item \code{upload_file()}: Uploads a file to the channel.
\item \code{download_file()}: Downloads a file from the channel.
}
}

\section{Initialization}{

Creating new objects of this class should be done via the \code{get_channel} and \code{list_channels} methods of the \link{ms_team} class. Calling the \code{new()} method for this class only constructs the R object; it does not call the Microsoft Graph API to retrieve or create the actual channel.
}

\examples{
\dontrun{

myteam <- team("my team")
myteam$list_channels()

chan <- myteam$get_channel()
chan$send_message("hello from R", attachments="myfile.csv")
chan$list_messages()

chan$upload_file("mydocument.docx")

}
}
\seealso{
\link{ms_team}, \link{ms_chat_message}

\href{https://docs.microsoft.com/en-us/graph/overview}{Microsoft Graph overview},
\href{https://docs.microsoft.com/en-us/graph/api/resources/teams-api-overview?view=graph-rest-1.0}{Microsoft Teams API reference}
}
