% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_plan.R
\docType{class}
\name{ms_plan}
\alias{ms_plan}
\title{Microsoft Planner Plan}
\format{
An R6 object of class \code{ms_plan}, inheriting from \code{ms_object}.
}
\description{
Class representing one plan withing a Microsoft Planner.
}
\details{
The plans belong to a group.
}
\section{Fields}{

\itemize{
\item \code{token}: The token used to authenticate with the Graph host.
\item \code{tenant}: The Azure Active Directory tenant for this drive.
\item \code{type}: always "plan" for plan object.
\item \code{properties}: The plan properties.
}
}

\section{Methods}{

\itemize{
\item \code{new(...)}: Initialize a new plan object. Do not call this directly; see 'Initialization' below.
\item \code{update(...)}: Update the plan metadata in Microsoft Graph.
\item \code{do_operation(...)}: Carry out an arbitrary operation on the plan
\item \code{sync_fields()}: Synchronise the R object with the plan metadata in Microsoft Graph.
\item \code{list_tasks(...)}: List the tasks under the specified plan.
}
}

\section{Initialization}{

Creating new objects of this class should be done via the \code{list_plans} methods of the \code{\link{az_group}} class.
Calling the \code{new()} method for this class only constructs the R object; it does not call the Microsoft Graph API to retrieve or create the actual plan.
}

\section{Planner operations}{

This class exposes methods for carrying out common operations on a plan.

\code{list_tasks()} lists the tasks under the plan.
}

\examples{
\dontrun{
}
}
\seealso{
\code{\link{ms_plan_task}}

\href{https://docs.microsoft.com/en-us/graph/overview}{Microsoft Graph overview},
\href{https://docs.microsoft.com/en-us/graph/api/resources/planner?view=graph-rest-1.0}{OneDrive API reference}
}
