% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_plan_task.R
\docType{class}
\name{ms_plan_task}
\alias{ms_plan_task}
\title{Microsoft Planner Plan Task}
\format{
An R6 object of class \code{ms_plan_task}, inheriting from \code{ms_object}.
}
\description{
Class representing a task within a plan of a Microsoft Planner.
}
\section{Fields}{

\itemize{
\item \code{token}: The token used to authenticate with the Graph host.
\item \code{tenant}: The Azure Active Directory tenant for this task.
\item \code{type}: always "plan_task" for plan task object.
\item \code{properties}: The task properties.
}
}

\section{Methods}{

\itemize{
\item \code{new(...)}: Initialize a new plan task object. Do not call this directly; see 'Initialization' below.
\item \code{update(...)}: Update the plan task metadata in Microsoft Graph.
\item \code{do_operation(...)}: Carry out an arbitrary operation on the plan task
\item \code{sync_fields()}: Synchronise the R object with the plan task metadata in Microsoft Graph.
}
}

\section{Initialization}{

Creating new objects of this class should be done via the \code{list_tasks} methods of the \code{\link{ms_plan}} class.
Calling the \code{new()} method for this class only constructs the R object; it does not call the Microsoft Graph API to retrieve or create the actual plan task.
}

\section{Plan task operations}{

This class exposes methods for carrying out common operations on a plan task.

\href{https://docs.microsoft.com/en-us/graph/overview}{Microsoft Graph overview},
\href{https://docs.microsoft.com/en-us/graph/api/resources/planner?view=graph-rest-1.0}{OneDrive API reference}
}

\examples{
\dontrun{
}
}
