% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_sharepoint_list.R
\docType{class}
\name{ms_sharepoint_list}
\alias{ms_sharepoint_list}
\title{Sharepoint list}
\format{
An R6 object of class \code{ms_sharepoint_list}, inheriting from \code{ms_object}.
}
\description{
Class representing a list in a SharePoint site.
}
\section{Fields}{

\itemize{
\item \code{token}: The token used to authenticate with the Graph host.
\item \code{tenant}: The Azure Active Directory tenant for the parent drive.
\item \code{type}: always "list" for a SharePoint list object.
\item \code{properties}: The item properties (metadata).
}
}

\section{Methods}{

\itemize{
\item \code{new(...)}: Initialize a new object. Do not call this directly; see 'Initialization' below.
\item \code{delete(confirm=TRUE)}: Delete this item. By default, ask for confirmation first.
\item \code{update(...)}: Update the item's properties in Microsoft Graph.
\item \code{do_operation(...)}: Carry out an arbitrary operation on the item.
\item \code{sync_fields()}: Synchronise the R object with the item metadata in Microsoft Graph.
\item \code{list_items(filter, select, all_metadata, pagesize)}: Queries the list and returns items as a data frame. See 'List querying below'.
\item \code{get_column_info()}: Return a data frame containing metadata on the columns (fields) in the list.
}
}

\section{Initialization}{

Creating new objects of this class should be done via the \code{get_list} method of the \link{ms_site} class. Calling the \code{new()} method for this class only constructs the R object; it does not call the Microsoft Graph API to retrieve or create the actual item.
}

\section{List querying}{

\code{list_items} supports the following arguments to customise results returned by the query.
\itemize{
\item \code{filter}: A string giving a logical expression to filter the rows to return. Note that column names used in the expression must be prefixed with \verb{fields/} to distinguish them from item metadata.
\item \code{select}: A string containing comma-separated column names to include in the returned data frame. If not supplied, includes all columns.
\item \code{all_metadata}: If TRUE, the returned data frame will contain extended metadata as separate columns, while the data fields will be in a nested data frame named \code{fields}.
\item \code{pagesize}: The number of results to return for each call to the REST endpoint. You can try reducing this argument below the default of 5000 if you are experiencing timeouts.
}

For more information, see \href{https://docs.microsoft.com/en-us/graph/query-parameters?view=graph-rest-1.0}{Use query parameters} at the Graph API reference. Due to current limitations in the Microsoft Graph REST API version 1.0, you can only read data from lists, not write to them.
}

\examples{
\dontrun{

site <- sharepoint_site("https://mycompany.sharepoint.com/sites/my-site-name")
lst <- site$get_list("mylist")

lst$get_column_info()

lst$list_items()
lst$list_items(filter="startswith(fields/firstname, 'John')", select="firstname,lastname")

}
}
\seealso{
\link{sharepoint_site}, \link{ms_site}

\href{https://docs.microsoft.com/en-us/graph/overview}{Microsoft Graph overview},
\href{https://docs.microsoft.com/en-us/graph/api/resources/sharepoint?view=graph-rest-1.0}{SharePoint sites API reference}
}
