% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_site.R
\docType{class}
\name{ms_site}
\alias{ms_site}
\title{Office 365 SharePoint site}
\format{
An R6 object of class \code{ms_site}, inheriting from \code{ms_object}.
}
\description{
Class representing a SharePoint site.
}
\section{Fields}{

\itemize{
\item \code{token}: The token used to authenticate with the Graph host.
\item \code{tenant}: The Azure Active Directory tenant for this site.
\item \code{type}: always "site" for a site object.
\item \code{properties}: The site properties.
}
}

\section{Methods}{

\itemize{
\item \code{new(...)}: Initialize a new site object. Do not call this directly; see 'Initialization' below.
\item \code{delete(confirm=TRUE)}: Delete a site. By default, ask for confirmation first.
\item \code{update(...)}: Update the site metadata in Microsoft Graph.
\item \code{do_operation(...)}: Carry out an arbitrary operation on the site.
\item \code{sync_fields()}: Synchronise the R object with the site metadata in Microsoft Graph.
\item \code{list_drives()}: List the drives (shared document libraries) associated with this site.
\item \code{get_drive(drive_id)}: Retrieve a shared document library for this site. If the ID is not specified, this returns the default document library.
\item \code{list_subsites()}: List the subsites of this site.
\item \code{get_lists()}: Returns the lists that are part of this site.
\item \code{get_list(list_name, list_id)}: Returns a specific list, either by name or ID.
}
}

\section{Initialization}{

Creating new objects of this class should be done via the \code{get_sharepoint_site} method of the \link{ms_graph} or \link{az_group} classes. Calling the \code{new()} method for this class only constructs the R object; it does not call the Microsoft Graph API to retrieve or create the actual site.
}

\examples{
\dontrun{

site <- get_sharepoint_site("My site")
site$list_drives()
site$get_drive()

}
}
\seealso{
\link{ms_graph}, \link{ms_drive}, \link{az_user}

\href{https://docs.microsoft.com/en-us/graph/overview}{Microsoft Graph overview},
\href{https://docs.microsoft.com/en-us/graph/api/resources/sharepoint?view=graph-rest-1.0}{SharePoint sites API reference}
}
