% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onedrive_login.R
\name{personal_onedrive}
\alias{personal_onedrive}
\alias{business_onedrive}
\title{OneDrive clients}
\usage{
personal_onedrive()

business_onedrive(tenant = "common")
}
\arguments{
\item{tenant}{For \code{business_onedrive}, the name of your Azure Active Directory (AAD) tenant. If not supplied, use the default tenant for your currently logged-in account.}
}
\value{
An object of class \code{ms_drive}.
}
\description{
OneDrive clients
}
\details{
These functions provide easy access to your OneDrive filesystem in the cloud. Use \code{personal_onedrive} to access the drive for your personal account, and \code{business_onedrive} to access the drive for your work or school account. For the latter, your organisation must have an appropriate Microsoft 365 license.

The default "common" tenant for \code{business_onedrive} attempts to detect your actual tenant from your currently logged-in account. This may not always succeed, for example if you have a personal account that is also a guest account in a tenant. In this case, supply the actual tenant name.
}
\examples{
\dontrun{

od <- personal_onedrive()
odb <- business_onedrive("myaadtenant")
odb$list_items()

}
}
\seealso{
\link{ms_drive}
}
