% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_channel.R
\docType{class}
\name{ms_channel}
\alias{ms_channel}
\title{Teams channel}
\format{
An R6 object of class \code{ms_channel}, inheriting from \code{ms_object}.
}
\description{
Class representing a Microsoft Teams channel.
}
\section{Fields}{

\itemize{
\item \code{token}: The token used to authenticate with the Graph host.
\item \code{tenant}: The Azure Active Directory tenant for the parent drive.
\item \code{type}: Always "channel" for a channel object
\item \code{team_id}: The ID of the parent team.
\item \code{properties}: The item properties (metadata).
}
}

\section{Methods}{

\itemize{
\item \code{new(...)}: Initialize a new object. Do not call this directly; see 'Initialization' below.
\item \code{delete(confirm=TRUE)}: Delete this channel. By default, ask for confirmation first.
\item \code{update(...)}: Update the channel's properties (metadata) in Microsoft Graph.
\item \code{do_operation(...)}: Carry out an arbitrary operation on the channel.
\item \code{sync_fields()}: Synchronise the R object with the channel metadata in Microsoft Graph.
\item \code{send_message(body, content_type, attachments)}: Sends a new message to the channel. See below.
\item \code{list_messages(n=50)}: Retrieves the messages in the channel. By default, this is limited to the 50 most recent messages; set the \code{n} argument to change this.
\item \code{get_message(message_id)}: Retrieves a specific message in the channel.
\item \code{delete_message(message_id, confirm=TRUE)}: Deletes a message. Currently the Graph API does not support deleting Teams messages, so this method is disabled.
\item \code{list_files()}: List the files for the channel. See \link{ms_drive} for the arguments available for this and the file upload/download methods.
\item \code{upload_file()}: Uploads a file to the channel.
\item \code{download_file()}: Downloads a file from the channel.
\item \code{get_folder()}: Retrieves the files folder for the channel, as a \link{ms_drive_item} object
}
}

\section{Initialization}{

Creating new objects of this class should be done via the \code{get_channel} and \code{list_channels} methods of the \link{ms_team} class. Calling the \code{new()} method for this class only constructs the R object; it does not call the Microsoft Graph API to retrieve or create the actual channel.
}

\section{Messaging}{

To send a message to a channel, use the \code{send_message()} method. This has arguments:
\itemize{
\item \code{body}: The body of the message. This should be a character vector, which will be concatenated into a single string with newline separators. The body can be either plain text or HTML formatted.
\item \code{content_type}: Either "text" (the default) or "html".
\item \code{attachments}: Optional vector of filenames.
}

Note that message attachments are actually uploaded to the channel's file listing (a directory in the team's primary shared document folder). Support for attachments is somewhat experimental, so if you want to be sure that it works, upload the file separately using the \code{upload_file()} method.
}

\examples{
\dontrun{

myteam <- get_team("my team")
myteam$list_channels()

chan <- myteam$get_channel()
chan$list_messages()
chan$send_message("hello from R")

# a multi-line message with an attachment
msg_text <- c(
    "message line 1",
    "message line 2",
    "message line 3"
)
chan$send_message(msg_text, attachments="myfile.csv")

chan$upload_file("mydocument.docx")

chan$list_files()

}
}
\seealso{
\link{ms_team}, \link{ms_drive}, \link{ms_chat_message}

\href{https://docs.microsoft.com/en-us/graph/overview}{Microsoft Graph overview},
\href{https://docs.microsoft.com/en-us/graph/api/resources/teams-api-overview?view=graph-rest-1.0}{Microsoft Teams API reference}
}
