% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_sharepoint_list.R
\docType{class}
\name{ms_sharepoint_list}
\alias{ms_sharepoint_list}
\title{Sharepoint list}
\format{
An R6 object of class \code{ms_sharepoint_list}, inheriting from \code{ms_object}.
}
\description{
Class representing a list in a SharePoint site.
}
\section{Fields}{

\itemize{
\item \code{token}: The token used to authenticate with the Graph host.
\item \code{tenant}: The Azure Active Directory tenant for the parent drive.
\item \code{type}: always "list" for a SharePoint list object.
\item \code{properties}: The item properties (metadata).
}
}

\section{Methods}{

\itemize{
\item \code{new(...)}: Initialize a new object. Do not call this directly; see 'Initialization' below.
\item \code{delete(confirm=TRUE)}: Delete this item. By default, ask for confirmation first.
\item \code{update(...)}: Update the item's properties in Microsoft Graph.
\item \code{do_operation(...)}: Carry out an arbitrary operation on the item.
\item \code{sync_fields()}: Synchronise the R object with the item metadata in Microsoft Graph.
\item \code{list_items(filter, select, include_metadata, pagesize)}: Queries the list and returns items as a data frame. See 'List querying below'.
\item \code{get_column_info()}: Return a data frame containing metadata on the columns (fields) in the list.
}
}

\section{Initialization}{

Creating new objects of this class should be done via the \code{get_list} method of the \link{ms_site} class. Calling the \code{new()} method for this class only constructs the R object; it does not call the Microsoft Graph API to retrieve or create the actual item.
}

\section{List querying}{

\code{list_items} supports the following arguments to limit the number of results returned by the query.
\itemize{
\item \code{filter}: A string giving a logical expression to filter the rows to return. Note that column names used in the expression must be prefixed with \verb{fields/} to distinguish them from item metadata.
\item \code{select}: A string containing comma-separated column names to filter the columns to return.
}

For more information, see \href{https://docs.microsoft.com/en-us/graph/query-parameters?view=graph-rest-1.0}{Use query parameters} on the Graph API reference.
}

\examples{
\dontrun{

site <- sharepoint_site("https://mycompany.sharepoint.com/sites/my-site-name")
lst <- site$get_list("mylist")

lst$get_column_info()

lst$list_items()
lst$list_items(filter="startswith(fields/firstname, 'John')", select="firstname,lastname")

}
}
\seealso{
\link{sharepoint_site}, \link{ms_site}

\href{https://docs.microsoft.com/en-us/graph/overview}{Microsoft Graph overview},
\href{https://docs.microsoft.com/en-us/graph/api/overview?view=graph-rest-1.0}{REST API reference}
}
